Run all tests in docker container
Previously, main_test go left out of docker image, so included that and
as a result had to change the test to point a test to Makefile instead of
LICENSE (its a error test) and also fixed the port usage to avoid conflict

Also override default port for HTTP listener in config package to 7050 to
avoid nondeterministic failure in acquiring port 8080; same for storage
package to 7040. The non-deterministic failure is more visible in container
build and hence fixed it as part of it.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>