Refactor how injection of repos and controllers happen
As number of controllers and repositories are getting close 4+ in each
category having that passed as factory parameters is making the code
more unreadable, so instead we went with factory objects to contain them
and use that for injection. In turn we also separated DB injector as pkg
private so that `main` uses its own injector making the abstraction more
stronger.

Also add a incomplete consumer repository. Add channel repo to router which
was mistakenly missed.

Change the DDL to create unique key as part of the table creation; works in
both SQLite and MySQL.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>