Fix queries to reflect correct column name spelling
Everywhere in the select queries we had `ID` instead if Id so fixed that.
How did the queries still work in sqlite is because column names are case
insensitive in that DB; but MySQL *nix servers are; so fixed them.

Also fix the test of adding ChannelController to router in unit test.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>