Use Mockery tool to generate mocks for testify/mock lib
This is done to avoid having to create and maintain the mocks manually. The
`make generate` will generate the mocks necessary for tests as of now.

Note these mocks are to be used outside of "storage" and "config" packages,
for in package use should be generated separately in _test files to avoid
cyclic import issues.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>