Add simple functions for better wrapping of transactions
For dispatcher we will need to have multi-insert+update SQL executed within
a single transaction, hence wanted a structure where we could leverage that
without any code duplication.

Also the fact that we do a `commit` even when there is no row update and
we are returning an error felt wrong, so reversed that to be a rollback.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>