Fix pagination cursor implementation
This fixes the issue #36. So instead of just using the ID as pagination
pointer it also adds created at time stamp at the point of page breaks to
traversal consistent with ordering.

So cursor is now a base64 encoded combination of ID and CreatedAt; so
change string coversation to ParseCursor and .String() usage instead. Also
add `createdAt` to all the pagination queries. This represents and another
challenge which now requires pagination to pass in query param for the date
to avoid formatting date in cross DB compatible manner. Also add DB indexes
for all list queries and add `id, createdAt` combo for recovery process
queries as well.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>