Introduce logging level
The issue seems to be with docker container handling a particular volume of
logs when the container starts up; it happens regardless of concurrency
level and it seemingly is happening only in the `ibroker` (broker WS)
container. So we introduced logging level to setup `Error` logging level
for integration test and that did the trick.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>