Add endpoints for messages, single message, dead jobs and their re-queuing
These are 4 new endpoints for listing messages against a channel, details
of a single message and its delivery, list of dead messages and a single
POST request to amend requeue the dead messages for a single try.

Reduce one of the bulk producer created as part of test fixture to not
change the assertions. Also add specific error for requeue not receiving
token as form param.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>