Make brittle tests more robust for slower CPUs
The dispatcher test waits for 200ms for 20 jobs to be requeued, which may
not be the case when all the tests are running with not enough CPU cores,
hence added a loop to wait for 4s for the condition to pass.

In the CLI Config test `OpenFileErrorInGetHash` is marked parallel and so
it would fail when it is running with `HashError` test in parallel, so
removed all parallel execution from this test.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>