Use Client VPN module to connect VPN to the VPC to access Elasticsearch
Since elasticsearch is in private subnet we needed VPN to access it. It
setting up the VPN we changed the default VPC security group to allow all
outbound connections, also created a specific security group for ES to
allow access to its 443 to make it accessible via VPC and VPN CIDRs.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>