Read the file as string first before creating a hash
While iterating on multiple fixes for file watcher, we noticed that for a
single write the notifier gets called twice and in both case the hash is
different and file write was attempted individually; evidently it's caused
by a truncation-file event.
Hence ignore any config file truncation; if a config file is truncated,
ignore that as config change as eventually it will have a configuration.
This is a limitation we are imposing on the file watch service; instead of
truncating, deletion of file is expected.

Also in main_test change the config files to be different to avoid any
notification conflict.

This change also seemingly causes dispatcher test to timeout and hence we
added 30s timeout to all test execution.

This seemingly fixes #52

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>