Setup release build script
The release build script is triggered via Tag Push. It will do -
1. Build package for the binary and attach to the Github Release
2. Build docker images and push to the repositories
3. Build Helm Package and upload it to Helm Repo

We did not introduce CHANGELOG in this release since its the first release;
before 0.2 or 0.1.x maintenance release we will introduce it.

Update the README with notes on using docker images for deployment.

Add developer notes on release management.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>