Refactor release CI process to use new Github Docker Registry
Github has launched the new Docker Registry which is better integrated
with Github Projects and the naming convention is less confusing. Also
the way GH Packages Docker publish was configured, that is deprecated in
the new version of build and push. So moving to the new registry is the
way forward.

Update the docker registry URL.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>