Fix channel update execution
The issue was a plain simple error in how values were updated for UPDATE
query and this is existing issue for all the other update queries as well.

The issue is, "preops" is what sets the value (in case of transaction
creation error to avoid updating the model), but the query arguments were
being set before it. So despite preops updating the value, query args were
getting the old values and hence no rows updated error.

The more worrying part was it was skipped in unit test entirely; the reason
for that is, unit test was using the returned value for change assertion.
To further confuse, SQLite seems to return row affected even if there isnt
any value change to the row hence it skipped our notice.

The fix is to pass pointers to the query arg to account for preops. Plus
in unit test instead of depending on sqlite row change to account for
actual data change, we load the entity using SELECT to ensure DB change
has actually happened.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>