Fix dispatching a message when there is no consumer for it
Previously when dispatching a message the assumption was there will always
be at least one job created for it; but that is an incorrect assumption in
case of a event bus; so we are gracefully handling it.

Without it the current behavior is no message will be dispatched until the
first consumer is attached, so the first consumer and first consumer alone
would receive all legacy messages, which is inconsistent behavior.

Hence the fix dispatches a message with no delivery job created; so any
new consumer attached to any channel at any sequence and any point of time
will have the DLQ to ingest legacy messages.

Fixes #104

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>