Make RetryCount to zero when jobs are pulled in from dlq
Previously when jobs are requeued from dlq by the consumer, their
previous retry count were not being reset to 0. The code flow is such that
for push job, each job will be tried once more and mark dead if fails
again. But for pull jobs, in the current flow without resetting the
retry count, the jobs will be put back in dead state just after they are
pulled from the dlq.