Update dependencies
In addition to updating dependencies, also update the GoLang version that the application is dependent on to 1.23. In that connection the base docker image is also updated, which now depends on Alpine 3.20.

On the way to updating dependency versions also made the tests
compatible with WSL2; the primary step was to skip tests that
depends on observing configuration file changes since that doesn't work in WSL2.