Add caching for channel, producer and consumer
Messages involved loading channel and producer; and loading job
involved loading consumer; so to avoid n*2 + 1 for loading messages we
added caching for channel, producer and consumer; this will be effective
largely when archiving thousands and millions of old messages and jobs.

This is wired through the dependency injection keeping it transparent to
rest of the stack via delegator pattern.