Add a single job requeue endpoint
A blank post would be suffice to requeue a dead job. The reason for adding a single job requeue is, if there is a consumer error and we want to test whether consumer is fixed, for push consumers, this endpoint will act like "step into" for testing a single job before requeue-ing the entire DLQ. Plus this could be used for more fine grain control on error recovery.

This is also part of resolution for #153