Add prune test flow in integration test
The setup is, there is a new container that is based of the webhook broker container, but also runs a nginx to expose the jsonl files written to it. The integration test first modifies the DB records directly to ensure the messages can be pruned and then tests whether at least a single file is written with jsonl. This ensures the basic pruning flow is working.

This is for #154